Add eye reminder script

+++ init.sh
@@ -10,5 +10,12 @@ ln -i -s $DIR/inputrc ~/.inputrc
 echo "Installing bin..."
 mkdir -p "$HOME/bin"
 for f in bin/*; do
-  ln -s "$PWD/$f" "$HOME/bin/"
+  if [ ! -f "$HOME/$f" ]; then
+    ln -s "$PWD/$f" "$HOME/bin/"
+  fi
 done
+
+if ! crontab -l | grep -q eye_reminder\\.sh; then
+  (crontab -l ; echo "*/20 * * * * \"${HOME}/bin/eye_reminder.sh\"") | crontab -
+fi
+
